put logic into the correct places
