+++ src/CribblyBackend.DataAccess/Repositories/TournamentRepository.cs
@@ -13,6 +13,7 @@ namespace CribblyBackend.DataAccess.Repositories
         Task<Tournament> GetNextTournament();
         Task<Tournament> Create(DateTime date);
         Task SetFlagValue(int tournamentId, string flagName, bool newVal);
+        Task<IEnumerable<Tournament>> GetTournamentsWithActiveFlag(string flagName);
     }
     public class TournamentRepository : ITournamentRepository
     {
@@ -55,28 +56,18 @@ namespace CribblyBackend.DataAccess.Repositories
             return tournaments.First();
         }
 
-        public async Task ChangeActiveStatus(int tournamentId, bool newVal)
+        public async Task<IEnumerable<Tournament>> GetTournamentsWithActiveFlag(string flagName)
         {
-            await SetFlagValue(tournamentId, nameof(Tournament.IsActive), newVal);
-        }
-        public async Task ChangeOpenForRegistrationStatus(int tournamentId, bool newVal)
-        {
-            await SetFlagValue(tournamentId, nameof(Tournament.IsOpenForRegistration), newVal);
+            // Note: this breaks the rule of using only parameterized query strings; however, the external world
+            // CANNOT control flagName here since flagName is passed by us and never from an external source
+            return await _connection.QueryAsync<Tournament>(
+                $@"SELECT * FROM Tournaments WHERE {flagName} = 1", // `true` doesn't exist in mysql; use 1
+                new { Name = flagName }
+            );
         }
 
         public async Task SetFlagValue(int tournamentId, string flagName, bool newVal)
         {
-            // Note: this breaks the rule of using only parameterized query strings; however, the external world
-            // CANNOT control flagName here since flagName is passed by us and never from an external source
-            var activeTournaments = (await _connection.QueryAsync<Tournament>(
-                $@"SELECT * FROM Tournaments WHERE {flagName} = 1", // `true` doesn't exist in mysql; use 1
-                new { Name = flagName }
-            )).ToList();
-            var (canSetValue, errMessage) = CanSetFlag(newVal, flagName, activeTournaments);
-            if (!canSetValue)
-            {
-                throw new Exception($"{errMessage} [attempted to change {flagName} status of {tournamentId}]");
-            }
             // Note: this breaks the rule of using only parameterized query strings; however, the external world
             // CANNOT control flagName here since flagName is passed by us and never from an external source
             await _connection.ExecuteAsync(
@@ -87,40 +78,6 @@ namespace CribblyBackend.DataAccess.Repositories
                 ",
                 new { Name = flagName, Value = newVal, Id = tournamentId }
                 );
-        }
-
-        private (bool, string) CanSetFlag(bool newVal, string flagName, List<Tournament> resultsWithFlagSet)
-        {
-            if (newVal)
-            {
-                if (resultsWithFlagSet.Count != 0)
-                {
-                    // can't set flag is a tournament already has the flag active
-                    var idList = string.Join(", ", resultsWithFlagSet.Select(t => t.Id.ToString()));
-                    return (false, $"Cannot set {flagName}; it is already set on tournament(s) {idList}");
-                }
-                // return early; nothing more to check if we're setting to true
-                return (true, "");
-            }
-            // Now we're unsetting a flag
-            if (resultsWithFlagSet.Count != 1)
-            {
-                // There should be exactly one tournament with the flag set if we're unsetting it
-                return (false, $"Cannot unset {flagName}; there is not exactly one tournament with this flag set");
-            }
-            // At this point we know there's only one result so can safely call First()
-            var result = resultsWithFlagSet.First();
-            var isActiveFlagInvalid =
-                flagName == nameof(Tournament.IsActive) && !result.IsActive;
-            var isOpenForRegistrationFlagInvalid =
-                flagName == nameof(Tournament.IsOpenForRegistration) && !result.IsOpenForRegistration;
-
-            if (isActiveFlagInvalid || isOpenForRegistrationFlagInvalid)
-            {
-                // The tournament we're unsetting the flag on should have it set
-                return (false, $"Cannot unset {flagName}; it isn't set on the specified tournament");
-            }
-            return (true, "");
         }
     }
 }
